Add 'stories' app with models, admin, and migrations
Added a new 'stories' app to the project. This includes the creation of the Story and Agent models, which will be used to handle data for narrative elements of the platform. Also included are the corresponding admin classes to manage these models from the Django admin interface. 'stories' app also added to the project's settings for inclusion in the installed apps.