Add handler for unknown commands in bot
An 'unknown_command_handler' was introduced to respond to undefined commands in the bot. This way, users get a response when they use an invalid command instead of no feedback. Additionally, a new static message 'unknown_command_md' was added, providing instruction to users on how to access a list of available commands. Lastly, reformatted 'dispatcher.py' to improve readability and eased future maintenance.