Add cover image support and remove old cover_prompt field
Changed the storytelling handlers in order to add visual support for stories. Now, if a story has a "cover_image_url", it will be shown in the response along with the text. Fixed some formatting issues as well.
Additionally, removed the "cover_prompt" field from the Story model since it was replaced by the "cover_image_url" field. Updated the Story admin page to reflect these changes.